Set auto or static IPv4 address.
